backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client==1.7.0
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
rank-bm25==0.2.2
scikit-learn==1.3.2
numpy==1.24.3